+++ libs/citygisoo/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added `ScrubLayer.extract_by_attribute()`.
 - Added `ScrubLayer.extract_by_expression()`.
 - Added `ScrubLayer.extract_by_aggregate_membership()`.
+- Added `ScrubLayer.difference_layer()`.
 
 ### Planned
 - Extend spatial join to support multiple predicates.